@@ -1,15 +1,18 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
-import mdx from 'vite-plugin-mdx'
-import md from './plugins/index'
-import Inspect from 'vite-plugin-inspect'
+import md from './plugins/index.ts'
+import inspect from 'vite-plugin-inspect'
+
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
-    react(),
-    md(),
-    Inspect(),
-    
+    inspect(),
+    react({
+      include: [/\.(ts|tsx)$/, /\.md$/],
+    }),
+    md({
+      include: /\.md$/
+    })
   ]
 })
